# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend Challenge",
        "version": "1.0.0",
        "title": "Turing/E-commerce",
        "contact": {
            "email": "codeninja0x01@gmail.com"
        }
    },
    "basePath": "/api",
...
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "bearer": {
            "type": "apiKey",
            "name": "USER-KEY",
            "in": "header"
        }
    },
    "host": "localhost:3000",
    "paths": {
...
        "/user": {
            "get": {
                "summary": "Get logged in user",
                "description": "Return current user.",
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    },
                    "401": {
                        "description": "Unauthorized User."
                    }
                },
                "tags": [
                    "user"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            },
            "post": {
                "summary": "Update user info",
                "description": "Update basic user info.",
                "parameters": [
                    {
                        "name": "name",
                        "required": true,
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "required": false,
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "dayPhone",
                        "required": true,
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "evePhone",
                        "required": false,
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "mobPhone",
                        "required": false,
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "creditCard",
                        "required": false,
                        "in": "formData",
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successfully operation.",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    },
                    "401": {
                        "description": "Unauthorized User."
                    }
                },
                "tags": [
                    "user"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ]
            }
        },
        "/user/address": {
            "post": {
                "summary": "Update user address",
                "description": "Update basic user address.",
                "parameters": [
                    {
                        "name": "address1",
                        "required": true,
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "address2",
                        "required": false,
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "city",
                        "required": true,
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "region",
                        "required": true,
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "country",
                        "required": true,
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "RegionId",
                        "required": true,
                        "in": "formData",
                        "type": "number"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successfully operation.",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    },
                    "401": {
                        "description": "Unauthorized User."
                    }
                },
                "tags": [
                    "user"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ]
            }
        },
...
    },
...
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "address1": {
                    "type": "string"
                },
                "address2": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "postalCode": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "dayPhone": {
                    "type": "string"
                },
                "evePhone": {
                    "type": "string"
                },
                "mobPhone": {
                    "type": "string"
                },
                "creditCard": {
                    "type": "string"
                }
            },
            "required": [
                "id",
                "name",
                "email",
                "address1",
                "address2",
                "city",
                "region",
                "postalCode",
                "country",
                "dayPhone",
                "evePhone",
                "mobPhone",
                "creditCard"
            ]
        },
...
    }
}